Adds functioning "editable" in handlebars template
Fixes the issue with "editable" not toggling between "true" and "false"
based on user ownership. Issue was with AJAX call for ViewAllProducts.
Updated the function to add an if statement to pass user token if
signed in and not pass it if user is not signed in. Fixed the issue.